feat: Update types, semantics, and type checker for atomic analysis
